Move more layer functions into map store
